Rework data layer, totally blow up the ui/tests lol
So I decided to come at this project from another angle, focusing on the data
first instead of the ui. I've put a lot of focus reworking the data layer
but as a result this totally breaks the UI and most of the tests (which
are view model tests).

I got impatient and instead of fixing everything first before merging
I just wanted to merge now and start making small commits getting things
working again, so the code is in a really shitty state right now but at
least all the tests are passing :D

Tags: refactor